chore: rename folder from local to utils
